update GitHub actions for nightly builds
This is a proposed update to allow GitHub Actions to run the nightly build in the same way that it currently runs the release build.  Assuming this works, the nightlybuild repository will no longer be needed.

This simply edits the config files for nightly builds in a similar way as the existing config files for release builds, so reviews and testing will be vital.