move g3_project_vertex in HudGaugeReticleTriangle::renderTriangle
In commit 09fad95070, the call to `g3_project_vertex` was moved outside the if() block.  In every other case where the vertex's `codes` field is compared with 0 (to see whether the vertex is on-screen or not) the call to `g3_project_vertex` appears after the check.  I therefore presume that the move was a mistake.