Create a copy of the object
Instead of returning the unique pointer.  This method is only used to get the value stored in the config and the reference will be destroyed pretty quickly after being called.  Sure, it is more efficient to use the pointer to get the info, but I don't think it's worth it to rewrite the methods here.

And it's not feasible to convert to shared pointer, since there are so many references to unique pointers in this code.  It's just better to just make a quick copy here.